pyproject.toml
[tool.poetry]
name = "our-league"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "our_league"}]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.23.2"
fastapi = "^0.101.1"
sqlalchemy = "^2.0.20"
psycopg2-binary = "^2.9.7"
pydantic-settings = "^2.0.3"
alembic = "^1.11.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
mypy = "^1.5.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
